Add Init() system and use it to catch a bug

+++ mtrandom/randomc.h
@@ -92,8 +92,8 @@
 
 // Define 32 bit signed and unsigned integers.
 // Change these definitions, if necessary, on 64 bit computers
-typedef   signed long int32;     
-typedef unsigned long uint32;     
+typedef   signed int int32;     
+typedef unsigned int uint32;     
 
 class TRandomMersenne {                // encapsulate random number generator
   #if 0
